+++ WildAnalyticsSDK/WildAnalyticsSDK/Sources/Device/Type.swift
@@ -1,0 +1,16 @@
+//
+//  Type.swift
+//  WildAnalyticsSDK
+//
+//  Copyright © 2025 Wildberries LLC. All rights reserved.
+//
+
+// swiftlint:disable identifier_name
+public enum Type: String {
+    case iPhone
+    case iPad
+    case iPod
+    case appleTV
+    case simulator
+    case unknown
+}