+++ WildAnalyticsSDK/WildAnalyticsSDK/WildAnalyticsSDK.h
@@ -1,0 +1,17 @@
+//
+//  WildAnalyticsSDK.h
+//  WildAnalyticsSDK
+//
+//
+
+#import <Foundation/Foundation.h>
+
+//! Project version number for WildAnalyticsSDK.
+FOUNDATION_EXPORT double WildAnalyticsSDKVersionNumber;
+
+//! Project version string for WildAnalyticsSDK.
+FOUNDATION_EXPORT const unsigned char WildAnalyticsSDKVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <WildAnalyticsSDK/PublicHeader.h>
+
+