+++ WildAnalyticsSDK/WildAnalyticsSDKTests/Utilities/DateTests.swift
@@ -1,0 +1,23 @@
+//
+//  Copyright © 2024 Wildberries LLC. All rights reserved.
+//
+
+import XCTest
+
+@testable import WildAnalyticsSDK
+
+final class DateTests: XCTestCase {
+
+    func testAsString() {
+        // given
+        let date = Date()
+        let dateFormatter = DateFormatter()
+        dateFormatter.locale = Locale(identifier: "en_US_POSIX")
+        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSXXXXX"
+        // when
+        let dateString = date.asString
+        // then
+        XCTAssertEqual(dateString, dateFormatter.string(from: date))
+    }
+
+}